create version 0.21
git-svn-id: https://odin-win.svn.sourceforge.net/svnroot/odin-win/trunk@24 b91d169b-bf28-450c-9b59-d70c078a1df8

+++ src/ODIN/SplitManagerCallback.h
@@ -30,5 +30,5 @@
 class ISplitManagerCallback {
 public:
   virtual void GetFileName(unsigned fileNo, std::wstring& fileName) = 0;
-  virtual int AskUserForMissingFile(LPCWSTR missingFileName, unsigned fileNo, std::wstring& newName) = 0;
+  virtual size_t AskUserForMissingFile(LPCWSTR missingFileName, unsigned fileNo, std::wstring& newName) = 0;
 };
