+++ testsrc/ODINTest/SplitFileTest.cpp
@@ -48,7 +48,7 @@ public:
     fileName += buf;
   }
 
-  virtual int AskUserForMissingFile(LPCWSTR missingFileName, unsigned fileNo, std::wstring& newName) {
+  virtual size_t AskUserForMissingFile(LPCWSTR missingFileName, unsigned fileNo, std::wstring& newName) {
     return IDCANCEL;
   }
 };
@@ -66,7 +66,7 @@ public:
     fileName += buf;
   }
 
-  virtual int AskUserForMissingFile(LPCWSTR missingFileName, unsigned fileNo, std::wstring& newName) {
+  virtual size_t AskUserForMissingFile(LPCWSTR missingFileName, unsigned fileNo, std::wstring& newName) {
     newName = fOrgName;
     GetFileName(fileNo, newName);
     return IDOK;
